Refactoring method names; add interfaces; add SequenceRecord
