+++ datavec-api/src/main/java/org/datavec/api/records/impl/SequenceRecord.java
@@ -14,30 +14,22 @@
  *  *    limitations under the License.
  */
 
-package org.datavec.api.records.reader;
+package org.datavec.api.records.impl;
 
-import org.datavec.api.berkeley.Pair;
-import org.datavec.api.records.SequenceRecord;
+import lombok.AllArgsConstructor;
+import lombok.Data;
 import org.datavec.api.records.metadata.RecordMetaData;
 import org.datavec.api.writable.Writable;
 
-import java.io.IOException;
 import java.util.List;
 
 /**
- * This is TEMPORARY interface to maintain forward compatibility until 0.7.0 release, at which point
- * the functionality here will be moved to SequenceRecordReader<br>
- *
- * SequenceRecordReaderMeta adds methods that provide both the record AND metadata for the record
- *
- * @author Alex
+ * Created by Alex on 20/09/2016.
  */
-public interface SequenceRecordReaderMeta extends SequenceRecordReader {
+@AllArgsConstructor @Data
+public class SequenceRecord implements org.datavec.api.records.SequenceRecord {
 
-    SequenceRecord nextSequence();
-
-    SequenceRecord loadFromMetaData(RecordMetaData recordMetaData) throws IOException;
-
-    List<SequenceRecord> loadFromMetaData(List<RecordMetaData> recordMetaDatas) throws IOException;
+    private List<List<Writable>> sequenceRecord;
+    private RecordMetaData metaData;
 
 }
