+++ datavec-api/src/main/java/org/datavec/api/records/reader/impl/csv/CSVRecordReader.java
@@ -18,12 +18,10 @@ package org.datavec.api.records.reader.impl.csv;
 
 
 
-import org.datavec.api.berkeley.Pair;
 import org.datavec.api.conf.Configuration;
 import org.datavec.api.records.Record;
 import org.datavec.api.records.metadata.RecordMetaData;
 import org.datavec.api.records.metadata.RecordMetaDataLine;
-import org.datavec.api.records.reader.RecordReader;
 import org.datavec.api.records.reader.RecordReaderMeta;
 import org.datavec.api.records.reader.impl.LineRecordReader;
 import org.datavec.api.writable.Text;
@@ -102,7 +100,7 @@ public class CSVRecordReader extends LineRecordReader implements RecordReaderMet
     }
 
     @Override
-    public Record nextMeta(){
+    public Record nextRecord(){
         List<Writable> next = next();
         URI uri = (locations == null || locations.length < 1 ? null : locations[0]);    //TODO: handle String splits etc
         RecordMetaData meta = new RecordMetaDataLine(this.currIndex-1, uri, CSVRecordReader.class); //-1 as line number has been incremented already...
@@ -110,12 +108,12 @@ public class CSVRecordReader extends LineRecordReader implements RecordReaderMet
     }
 
     @Override
-    public Record loadFromMeta(RecordMetaData recordMetaData) throws IOException {
-        return loadFromMeta(Collections.singletonList(recordMetaData)).get(0);
+    public Record loadFromMetaData(RecordMetaData recordMetaData) throws IOException {
+        return loadFromMetaData(Collections.singletonList(recordMetaData)).get(0);
     }
 
     @Override
-    public List<Record> loadFromMeta(List<RecordMetaData> recordMetaDatas) throws IOException {
+    public List<Record> loadFromMetaData(List<RecordMetaData> recordMetaDatas) throws IOException {
         List<Record> list = super.loadFromMeta(recordMetaDatas);
 
         for(Record r : list ){
